# bytes, strings: optimize Cut for single-byte separators