--- a/src/strings/strings_test.go
+++ b/src/strings/strings_test.go
@@ -2071,3 +2071,31 @@ func BenchmarkReplaceAll(b *testing.B) {
 		stringSink = ReplaceAll("banana", "a", "<>")
 	}
 }
+
+func BenchmarkCut(b *testing.B) {
+	b.ReportAllocs()
+
+	for _, skip := range [...]int{2, 4, 8, 16, 32, 64} {
+		s := Repeat(Repeat(" ", skip)+"aa"+Repeat(" ", skip), 1<<16/skip)
+		b.Run(fmt.Sprintf("Cut-One/%d", skip), func(b *testing.B) {
+			for range b.N {
+				_, _, _ = Cut(s, "a")
+			}
+		})
+		b.Run(fmt.Sprintf("Cut-Two/%d", skip), func(b *testing.B) {
+			for range b.N {
+				_, _, _ = Cut(s, "aa")
+			}
+		})
+		b.Run(fmt.Sprintf("Cut-One-Nil/%d", skip), func(b *testing.B) {
+			for range b.N {
+				_, _, _ = Cut(s, "c")
+			}
+		})
+		b.Run(fmt.Sprintf("Cut-Two-Nil/%d", skip), func(b *testing.B) {
+			for range b.N {
+				_, _, _ = Cut(s, "cc")
+			}
+		})
+	}
+}